submitQuestion in useChatState and renderConversation components broken down
